+++ lib/post-types/pdf-post-type.php
@@ -1,5 +1,6 @@
 <?php
 
+/* Add PDF Post Type */
 
 add_action('init', 'masked_pdf_post_type');